Creation de l'activité detail anime pour afficher les details + fix du bug error 500 + layout affichage grille anime depuis les genres fait

+++ app/src/main/java/com/example/animelist/Model/AnimeDetail.java
@@ -17,7 +17,7 @@ public class AnimeDetail {
     @SerializedName("episodes")
     private int episodes;
     @SerializedName("score")
-    private long score;
+    private double score;
     @SerializedName("synopsis")
     private String synopsis;
     @SerializedName("premiered")
@@ -61,7 +61,7 @@ public class AnimeDetail {
         this.episodes = episodes;
     }
 
-    public long getScore() {
+    public double getScore() {
         return score;
     }
